Copy the input arguments for async functions locally so that we are dependent on the caller maintaining the lifetime until the async function completes.  Also a bunch of cleanup working and updating to the latest XSAPI nuget package. (#170)

+++ CSharpSource/Source/api/Privacy/MultiplePermissionsCheckResult.cs
@@ -39,6 +39,6 @@ namespace Microsoft.Xbox.Services.Privacy
     {
         public IntPtr xboxUserId;
         public IntPtr items;
-        public UInt64 itemsCount;
+        public UInt32 itemsCount;
     }
 }
